% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_data.R
\docType{data}
\name{fossil_data}
\alias{fossil_data}
\title{Fossil Data}
\format{
A data frame with 512 rows and 12 columns:
\describe{
\item{Age_BP}{Calibrated age in years BP}
\item{Si}{Si data}
\item{S}{S data}
\item{K}{K data}
\item{Ca}{Ca data}
\item{Ti}{Ti data}
\item{V}{V data}
\item{Mn}{Mn data}
\item{Fe}{Fe data}
\item{Rb}{Rb data}
\item{Sr}{Sr data}
\item{Zr}{Zr data}
}
}
\source{
Boyall et al (2024, JQS)
}
\usage{
fossil_data
}
\description{
A dataset containing fossil records for the reconstruction model.
}
\keyword{datasets}
