% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern_data.R
\docType{data}
\name{modern_data}
\alias{modern_data}
\title{Modern Data}
\format{
A data frame with 199 rows and 13 columns:
\describe{
\item{Age_BP}{Calibrated age in years BP}
\item{Annual}{Annual temperature anomalies}
\item{Si}{Si data}
\item{S}{S data}
\item{K}{K data}
\item{Ca}{Ca data}
\item{Ti}{Ti data}
\item{V}{V data}
\item{Mn}{Mn data}
\item{Fe}{Fe data}
\item{Rb}{Rb data}
\item{Sr}{Sr data}
\item{Zr}{Zr data}
}
}
\source{
Boyall et al (2024, JQS)
}
\usage{
modern_data
}
\description{
A dataset containing modern records for the calibration model.
}
\keyword{datasets}
