% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBIDO_apply.R
\name{SCUBIDO_apply}
\alias{SCUBIDO_apply}
\title{Apply the relationship between modern climate and XRF core scanning data and apply to the fossil data}
\usage{
SCUBIDO_apply(calibration, temp_grid, plot_graph = FALSE, print = FALSE)
}
\arguments{
\item{temp_grid}{a sequence range between two values for the model to pick the temperature from. if working with anomalies use seq(-3, 3, length = 50)}

\item{plot_graph}{a function to plot the output of the SCUBIDO_apply function}

\item{print}{a function which prints out the output}

\item{sorted}{the sorted data set coming from the first function}

\item{calibration_data}{is the output list from the previous SCUBIDO_cal function.}
}
\value{
a data set containing the output
}
\description{
The SCUBIDO_apply function takes what is learnt in the modern observational
period in \code{\link{SCUBIDO_cal}} and applies this relationship to the fossil data.
}
\examples{
\dontrun{
temp_grid <- seq(3,3, length = 50)
MDP <- SCUBIDO_apply(calibration_data, temp_grid, plot_graph = T, print = F)
}


}
