% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBIDO_cal.sorted.R
\name{SCUBIDO_cal.sorted}
\alias{SCUBIDO_cal.sorted}
\title{Calculate the modern relationship between the proxy and climate (sorted method)}
\usage{
\method{SCUBIDO_cal}{sorted}(sorted, plot = TRUE, summary = TRUE)
}
\arguments{
\item{sorted}{the modern dataset saved after using the \code{\link{SCUBIDO_input}} function}

\item{plot}{returns a plot of the relationship between the modern XRF elements and climate}

\item{summary}{returns a printed summary of the output of the calibration model}
}
\description{
This method is specifically for datasets that have been sorted using
the \code{\link{SCUBIDO_input}} function.
}
\examples{
\dontrun{
SCUBIDO_cal(sorted,  plot = TRUE, summary = TRUE)
}

\dontrun{
SCUBIDO_cal(x, plot = TRUE, summary = TRUE)
}
}
