% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBIDO_input.R
\name{SCUBIDO_input}
\alias{SCUBIDO_input}
\title{A function which sorts the modern and fossil proxy data sets}
\usage{
SCUBIDO_input(modern_data, fossil_data)
}
\arguments{
\item{modern_data}{a data set containing age in BCE/CE in the first column, a climate data set in the second column and then the third column thereafter contain all of the XRF elements.}

\item{fossil_data}{a data set containing age in BCE/CE in the first column,the second column and thereafter contain all of the XRF elements used in the modern data}
}
\value{
list of transformed proxy data to be used in the models and a validation dataset
}
\description{
This function takes the modern data set (modern_data) containing age in BCE/CE in the first
column, a modern observational climate timeseries in the second column
and then a final array of columns represeting the different xrf elements.
The second data set is the fossil data (fossil_data) and this take the same
approach as the modern, but contains one less column as there is no
climate data included.
This function then takes these two data sets and rescales the fossil proxy
data to be consistent with the modern xrf data. It then returns a list
containing the data required for the next stages of the proxy modelling.
}
\examples{
# A simple example containing 50 years of modern data and 150 years
# of fossil data

\dontrun{
data(modern_data, fossil_data)
SCUBIDO_input(modern_data, fossil_data)
}
}
