% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBIDO_reconstruct.R
\name{SCUBIDO_reconstruct.MDPs}
\alias{SCUBIDO_reconstruct.MDPs}
\title{Reconstruction of climate}
\usage{
SCUBIDO_reconstruct.MDPs(MDPs, time_grid, plot_graph = FALSE, print = FALSE)
}
\arguments{
\item{MDPs}{the MDP list created in the apply function}

\item{time_grid}{The maximum range of age in BCE/CE from your modern data set to your fossil data set e.g. range(fossil_data$age, modern_data$age) and then sequence by wanted resolution e.g. 10 years. see example.}

\item{plot}{A plot of the final reconstruction}
}
\value{
a full reconstruction of climate given the proxt data
}
\description{
SCUBIDO_reconstruct is a function which sorts the previous analysis from the
calibration period and the marginal data posteriors and applies a random
walk
}
\examples{
\dontrun{
time_grid <- seq(-8331.625, 1932.000, by = 10)
SCUBIDO_reconstruct(time_grid, plot = TRUE)
}

}
