% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBIDO_reconstruct.R
\name{SCUBIDO_reconstruct}
\alias{SCUBIDO_reconstruct}
\title{Reconstruction of climate}
\usage{
SCUBIDO_reconstruct(calibration_data, plot_graph = FALSE)
}
\arguments{
\item{calibration_data}{is the output list from the SCUBIDO_cal function.}

\item{plot_graph}{A flag to plot the output of the reconstruction.}
}
\value{
a full reconstruction of climate given the proxy data.
}
\description{
SCUBIDO_reconstruct is a function which sorts the previous analysis from the
calibration period and then applies what is learnt on the fossil_data.
This first creates marginal data posteriors to transform the multivariate
proxy data into a single estimation of climate based on each time point
of the XRF data. A random walk is then applied to join each of these layers
and reconstruct climate through time and quantifies 95\% and 50\% confidence
intervals.
Note that this function can take many hours to run depending on how many data
points in the XRF data, a dataset containing approximately 60,000 time points
takes around 12 hours on a standard computer.
}
\examples{
\dontrun{
SCUBIDO_reconstruct(calibration_data, plot_graph = TRUE)
}
}
