% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_data.R
\docType{data}
\name{fossil_data}
\alias{fossil_data}
\title{Fossil Data}
\format{
A data frame with 570 rows and 14 columns:
\describe{
\item{Age.CE}{Common Era Years}
\item{Al.clr}{Al data}
\item{Si.clr}{Si data}
\item{K.clr}{K data}
\item{Ca.clr}{Ca data}
\item{Ti.clr}{Ti data}
\item{V.clr}{V data}
\item{Cr.clr}{Cr data}
\item{Mn.clr}{Mn data}
\item{Fe.clr}{Fe data}
\item{Zn.clr}{Zn data}
\item{Rb.clr}{Rb data}
\item{Sr.clr}{Sr data}
\item{Zr.clr}{Zr data}
}
}
\usage{
fossil_data
}
\description{
A dataset containing fossil records for the reconstruction model.
}
\keyword{datasets}
